requirements.txt
flask==3.0.2
python-dotenv